requirements.txt
feedparser==6.0.10
pandas==1.5.2
python-dotenv==1.0.0
